Add files via upload
To write a program that prompts the user to enter a quantity and a cost. The values are read into an integer named quantity and a float named unit price. Define the variables and use only one statement to read the values. After reading the values, skip one line and print each value, with an appropriate name, on a separate line.