Add files via upload
Write a program to create a customer's bill for a company. The company sells only five different products: TV, VCR, remote controller, CD player and tape recorder. The unit prices are Rs. 400, 220, 35.20, 300, and 150, respt after an 8.25% sale taxectively. 
The program must read from the keyboard the quantity of each piece of equipment purchased. it then calculates the cost of each item, the subtotal and the total cost.
The input data consists of a set of integers representing the quantities of each item sold. These integers must be input into the program in a user-friendly way; That is, the program must prompt the user for each quantity as shown below. The numbers in the shows the general questions asked to the user:

How many TV's are sold? 
3 
How many VCR's were sold?
5 
How many remote controllers were sold?
1 
How many CDs were sold? 
2
How many tape recorders were sold?
4

Based on it generate a particular Bill for the customer.