Add files via upload
A bookshop maintains the inventory of books that are being sold at the shop. The list includes details such as author, title, price, publisher and stock positions. Whenever a customer wants a book, the salesperson input the title and author and the system searches the list and displays whether it is valid or not. If it is not,an appropriate message is displayed.If it is, then the system displays the book details and requests for the number of copies required.If the required copies are available, the total cost of the requested copies is displayed; otherwise the message required copies not in stock is displayed.